moved interface Authentication Constant Docs
