+++ src/main/java/com/alkemy/ong/dto/UserAuthenticatedResponseDto.java
@@ -1,6 +1,6 @@
 package com.alkemy.ong.dto;
 
-import com.alkemy.ong.controller.docs.AuthenticationConstantDocs;
+import com.alkemy.ong.util.docs.AuthenticationConstantDocs;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.*;
