+++ src/main/java/com/alkemy/ong/security/dto/LoggedUserDto.java
@@ -1,6 +1,6 @@
 package com.alkemy.ong.security.dto;
 
-import com.alkemy.ong.controller.docs.AuthenticationConstantDocs;
+import com.alkemy.ong.util.docs.AuthenticationConstantDocs;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
