+++ src/main/java/com/alkemy/ong/util/docs/AuthenticationConstantDocs.java
@@ -1,4 +1,4 @@
-package com.alkemy.ong.controller.docs;
+package com.alkemy.ong.util.docs;
 
 public interface AuthenticationConstantDocs {
